refactor: Move getFilesPath to helper
